+++ src/solvers/dg_gpu.jl
@@ -12,6 +12,11 @@ using Atomix: @atomic
     return nothing
 end
 
+@inline function get_node_vars_local(flux_local, ::Val{NVARIABLES},
+                                     indices...) where {NVARIABLES}
+    return SVector(ntuple(v -> (@inbounds flux_local[v, indices...]), Val(NVARIABLES)))
+end
+
 function calc_volume_integral!(backend::Backend, du, u, mesh,
                                have_nonconservative_terms, equations,
                                volume_integral, dg::DGSEM, cache)