+++ test/test_amdgpu_3d.jl
@@ -230,3 +230,65 @@ end
     @test Trixi.storage_type(semi.cache.boundaries) === ROCArray
     @test Trixi.storage_type(semi.cache.mortars) === ROCArray
 end
+
+@testitem "AMDGPU 3D: elixir_euler_source_terms_nonperiodic.jl HalfSweep vs. FullSweep / AMDGPU" setup=[
+    Setup,
+    AMDGPU3DExamples
+] tags=[:AMDGPU] begin
+    # Using AMDGPU inside the testitem since otherwise the bindings are hidden by the anonymous modules
+    using AMDGPU
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR, "elixir_euler_source_terms_nonperiodic.jl"),
+                  volume_integral = VolumeIntegralFluxDifferencing(flux_kennedy_gruber),
+                  tspan = (0.0, 0.5),
+                  real_type = Float32,
+                  storage_type = ROCArray,
+                  flux_differencing_kernel = HalfSweep())
+    @test ode.p.cache.flux_differencing_kernel === HalfSweep()
+    @test Trixi.storage_type(ode.p.cache.elements) === ROCArray
+    u_half_sweep = Array(sol.u[end])
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR, "elixir_euler_source_terms_nonperiodic.jl"),
+                  volume_integral = VolumeIntegralFluxDifferencing(flux_kennedy_gruber),
+                  tspan = (0.0, 0.5),
+                  real_type = Float32,
+                  storage_type = ROCArray,
+                  flux_differencing_kernel = FullSweep())
+    @test ode.p.cache.flux_differencing_kernel === FullSweep()
+    u_full_sweep = Array(sol.u[end])
+
+    @test u_half_sweep ≈ u_full_sweep
+end
+
+@testitem "AMDGPU 3D: elixir_mhd_alfven_wave_combined_fluxes_nonperiodic.jl HalfSweep vs. FullSweep / AMDGPU" setup=[
+    Setup,
+    AMDGPU3DExamples
+] tags=[:AMDGPU] begin
+    # Using AMDGPU inside the testitem since otherwise the bindings are hidden by the anonymous modules
+    using AMDGPU
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR,
+                           "elixir_mhd_alfven_wave_combined_fluxes_nonperiodic.jl"),
+                  tspan = (0.0, 0.1),
+                  real_type = Float32,
+                  storage_type = ROCArray,
+                  flux_differencing_kernel = HalfSweep())
+    @test ode.p.cache.flux_differencing_kernel === HalfSweep()
+    @test Trixi.storage_type(ode.p.cache.elements) === ROCArray
+    u_half_sweep = Array(sol.u[end])
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR,
+                           "elixir_mhd_alfven_wave_combined_fluxes_nonperiodic.jl"),
+                  tspan = (0.0, 0.1),
+                  real_type = Float32,
+                  storage_type = ROCArray,
+                  flux_differencing_kernel = FullSweep())
+    @test ode.p.cache.flux_differencing_kernel === FullSweep()
+    u_full_sweep = Array(sol.u[end])
+
+    @test u_half_sweep ≈ u_full_sweep
+end