+++ test/test_cuda_3d.jl
@@ -227,3 +227,65 @@ end
     @test Trixi.storage_type(semi.cache.boundaries) === CuArray
     @test Trixi.storage_type(semi.cache.mortars) === CuArray
 end
+
+@testitem "CUDA 3D: elixir_euler_source_terms_nonperiodic.jl HalfSweep vs. FullSweep / CUDA" setup=[
+    Setup,
+    CUDA3DExamples
+] tags=[:CUDA] begin
+    # Using CUDA inside the testitem since otherwise the bindings are hidden by the anonymous modules
+    using CUDA
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR, "elixir_euler_source_terms_nonperiodic.jl"),
+                  volume_integral = VolumeIntegralFluxDifferencing(flux_kennedy_gruber),
+                  tspan = (0.0, 0.5),
+                  real_type = Float32,
+                  storage_type = CuArray,
+                  flux_differencing_kernel = HalfSweep())
+    @test ode.p.cache.flux_differencing_kernel === HalfSweep()
+    @test Trixi.storage_type(ode.p.cache.elements) === CuArray
+    u_half_sweep = Array(sol.u[end])
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR, "elixir_euler_source_terms_nonperiodic.jl"),
+                  volume_integral = VolumeIntegralFluxDifferencing(flux_kennedy_gruber),
+                  tspan = (0.0, 0.5),
+                  real_type = Float32,
+                  storage_type = CuArray,
+                  flux_differencing_kernel = FullSweep())
+    @test ode.p.cache.flux_differencing_kernel === FullSweep()
+    u_full_sweep = Array(sol.u[end])
+
+    @test u_half_sweep ≈ u_full_sweep
+end
+
+@testitem "CUDA 3D: elixir_mhd_alfven_wave_combined_fluxes_nonperiodic.jl HalfSweep vs. FullSweep / CUDA" setup=[
+    Setup,
+    CUDA3DExamples
+] tags=[:CUDA] begin
+    # Using CUDA inside the testitem since otherwise the bindings are hidden by the anonymous modules
+    using CUDA
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR,
+                           "elixir_mhd_alfven_wave_combined_fluxes_nonperiodic.jl"),
+                  tspan = (0.0, 0.1),
+                  real_type = Float32,
+                  storage_type = CuArray,
+                  flux_differencing_kernel = HalfSweep())
+    @test ode.p.cache.flux_differencing_kernel === HalfSweep()
+    @test Trixi.storage_type(ode.p.cache.elements) === CuArray
+    u_half_sweep = Array(sol.u[end])
+
+    trixi_include(@__MODULE__,
+                  joinpath(EXAMPLES_DIR,
+                           "elixir_mhd_alfven_wave_combined_fluxes_nonperiodic.jl"),
+                  tspan = (0.0, 0.1),
+                  real_type = Float32,
+                  storage_type = CuArray,
+                  flux_differencing_kernel = FullSweep())
+    @test ode.p.cache.flux_differencing_kernel === FullSweep()
+    u_full_sweep = Array(sol.u[end])
+
+    @test u_half_sweep ≈ u_full_sweep
+end